pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "RAG-based LLM-powered document search"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.4",
    "httpx",
    "pymupdf4llm",
    "pyyaml",
    "qdrant-client",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[project.scripts]
app = "app:main"
app-ingest = "app.ingest:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
python_files = ["test_*.py", "*_test.py", "*_tests.py"]
markers = [
    "integration: tests requiring running Qdrant and Ollama services",
]

[build-system]
requires = ["uv_build>=0.9.24,<0.10.0"]
build-backend = "uv_build"
